+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 
 * Add support for `bank_*` and `contact_*` attributes in generic-transfer model
+* Add `state` attribute to generic-transfer model
 
 ## v2.0.2